Added configuration parameter to allow the parsing of non-compliant OSC addresses
Added a parameter to the top level module called `OSC_DGRAM_STARTS_WITH` which is used to determine whether a datagram is a valid OSC message. The check originally checked to see if the message (and thus address) started with a `/`. Exposing this as a configuration option allows this to be set to something else, which is needed for controlling devices that are non-compliant with the OSC address specification (for example the Behringer X32 audio mixing desk series)